pyproject.toml
[project]
name = "mineshield"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "joblib>=1.4.2",
    "matplotlib>=3.10.1",
    "mlxtend>=0.23.4",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pyqt5>=5.15",
    "pyqt5-tools>=5.15.9.3.3",
    "pywin32>=310",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "umap-learn>=0.5.7",
    "win32mica>=3.0",
]
